% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{content.clust}
\alias{content.clust}
\alias{content.clust-package}
\title{content.clust}
\description{
a set of methods for cell clustering in scRNAseq data, that use quantities' of cells that express particular gene relationships with each other
}
\author{
RedSnail
}
